Shadows
`box-shadow` and `text-shadow` take almost the same arguments, expect
that the former has a numeric argument for a "spread" value, and it can
be `inset`.

For this reason, values of `text-shadow` can be reused as values of
`box-shadow`, but not viceversa, so I've put `-box-` in the middle of
the name of variales holding values for `box-shadow` but nothing in the
names of variables holding values for `text-shadow`.