`<nav>` functionality via jQuery
The tab-based nav works via jQuery now.

Two callbacks are used:

  - When the hash change in the URL, all s are hidden, then only the one pointed by the hash is shown, and the corresponding checkbox is checked;
  - when a radio button is clicked, the hash is changed accordingly, thus triggering the other callback too.

To start up the hash is set according to the radio button which is  by default, and the  event is triggered.

The advantange of this hash-based navigation is that it automatically saves each tab selection as an entry in the browser history.

I am not sure how I will be able to do something similar for sub-tabs or filters, or whatever, considering that at most one hash exists in the URL (that is what I have understood). Probably I'll have to use queries.