Fix radial diagram bug
Long story short, one of the rectangles of the skills was so skewed that
the outward corner (which is in the transparent half of the rectangle)
was protruding beyond the `<html>` border, resulting in an horizontal
scroll bar that looked misterious.

I've hidden the overflow to prevent that.

On the other hand, I had to change also the position of the language
logos, pulling them inside the canvas, rather than having them outside,
otherwise they are cut out by the `overflow: hidden` rule